+++ uvm/api/com/untangle/uvm/SessionMatcher.java
@@ -16,5 +16,5 @@ public interface SessionMatcher
     /**
      * Alternate session match test
      */
-    boolean isMatch( Long policyId, short protocol, int clientIntf, int serverIntf, InetAddress clientAddr, InetAddress serverAddr, int clientPort, int serverPort, Map<String,Object> attachments );
+    boolean isMatch( Integer policyId, short protocol, int clientIntf, int serverIntf, InetAddress clientAddr, InetAddress serverAddr, int clientPort, int serverPort, Map<String,Object> attachments );
 }
